fix: fetch repo assets at build
