+++ CommunityCreations.html
@@ -272,7 +272,7 @@
                 class="carousel-slide flex-none p-4 flex flex-col sm:flex-row items-center justify-center gap-4"
               >
                 <img
-                  src="img/astro image.jpg"
+                  src="img/astro-image.png"
                   alt="Real-life print"
                   loading="lazy"
                   class="w-full sm:w-1/2 h-64 object-cover rounded-xl"
