+++ package.json
@@ -37,7 +37,7 @@
     "typecheck": "tsc --noEmit",
     "validate-env": "bash scripts/validate-env.sh",
     "setup": "bash scripts/setup.sh",
-    "prebuild": "node scripts/fetch-boombox.mjs",
+    "prebuild": "node scripts/fetch-assets.mjs",
     "build": "npm ci --no-audit --no-fund --prefix frontend && npm --prefix frontend run build",
     "start:ci": "npm --prefix frontend run start:ci",
     "preview": "vite preview --port 4173",
